# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "adhocly-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local adhocly API"
    }
  ],
  "paths": {
...
    "/api/reminders/{id}": {
      "get": {
        "tags": [
          "reminders"
        ],
        "operationId": "getReminder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Reminder id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reminder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reminder"
                }
              }
            }
          },
          "404": {
            "description": "Reminder not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "operationId": "getTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tasks"
        ],
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid task",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted task"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomReminder": {
        "type": "object",
        "required": [
          "id",
          "kind"
        ],
        "properties": {
          "at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock time; required only for custom-date reminders."
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ReminderKind"
          },
          "offsetMinutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Legacy presets; omit when using offsetUnit and offsetValue."
          },
          "offsetUnit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReminderUnit"
              }
            ]
          },
          "offsetValue": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Whole units before the anchor, from 1 through 999."
          },
          "urgent": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Reminder": {
        "type": "object",
        "required": [
          "remindAt",
          "taskId"
        ],
        "properties": {
          "remindAt": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          }
        }
      },
      "ReminderKind": {
        "type": "string",
        "enum": [
          "custom",
          "planned",
          "due"
        ]
      },
      "ReminderUnit": {
        "type": "string",
        "enum": [
          "minutes",
          "hours",
          "days",
          "weeks",
          "months"
        ]
      },
...
      "Task": {
        "type": "object",
        "required": [
          "id",
          "title",
          "projectId",
          "project",
          "createdAt",
          "updatedAt",
          "completed"
        ],
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "dueOn": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock date and time, `YYYY-MM-DDTHH:MM`."
          },
          "id": {
            "type": "string"
          },
          "plannedFor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock date and time, `YYYY-MM-DDTHH:MM`."
          },
          "project": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "reminders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomReminder"
            }
          },
          "repeatWeekday": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "statusId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "TaskInput": {
        "type": "object",
        "required": [
          "id",
          "title",
          "projectId",
          "completed"
        ],
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "details": {
            "type": [
              "string",
              "null"
            ],
            "description": "Omission preserves details; an empty string clears them."
          },
          "dueOn": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock date and time, `YYYY-MM-DDTHH:MM`."
          },
          "id": {
            "type": "string"
          },
          "plannedFor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock date and time, `YYYY-MM-DDTHH:MM`."
          },
          "projectId": {
            "type": "string"
          },
          "reminders": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CustomReminder"
            },
            "description": "Omission preserves reminders; an empty list removes all custom reminders."
          },
          "repeatWeekday": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "statusId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Omission preserves an existing non-complete status on updates from older clients."
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}